removed global lock on snmp trap configuration; minor fixes

+++ src/client/windows/nxcon/ActionSelDlg.cpp
@@ -112,7 +112,7 @@ void CActionSelDlg::OnOK()
       iItem = m_wndListCtrl.GetNextItem(-1, LVIS_SELECTED);
       for(i = 0; iItem != -1; i++)
       {
-         m_pdwActionList[i] = m_wndListCtrl.GetItemData(iItem);
+         m_pdwActionList[i] = (DWORD)m_wndListCtrl.GetItemData(iItem);
          iItem = m_wndListCtrl.GetNextItem(iItem, LVIS_SELECTED);
       }
 	   CDialog::OnOK();
